+++ dashboards/customized/influxdb_v3/plugin/dates.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from datetime import datetime, timezone
